Attempting to fix font scaling in Linux
Phos reported to me that there was a recurring problem on the Linux system, where some fonts provided different values than what was shown on the Aegisub screen. He saw that the SubInspector module provided this data accurately, so he advised me to try using these values to make this attempt at a fix, which is exactly what I did.
The truth is that the values of this correction are not accurate at all, it is not a definitive solution.
I don't know if it is an Aegisub error, but the fact is that it doesn't provide the correct font metrics, which ends up causing this problem.